Created Slider Component
Moved slider to seperate file to build component, then render app, adding functionality and flexability.
